fmt script/

+++ script/DeployLib.sol
@@ -7,10 +7,11 @@ import {InitSetAdmin} from "@mc-std/functions/protected/InitSetAdmin.sol";
 
 library DeployLib {
     using DeployLib for MCDevKit;
+
     string internal constant BUNDLE_NAME = "Std";
 
-    function deployStd(MCDevKit storage mc, address admin) internal returns(address) {
-        for (uint i; i < mc.std.all.functions.length; ++i) {
+    function deployStd(MCDevKit storage mc, address admin) internal returns (address) {
+        for (uint256 i; i < mc.std.all.functions.length; ++i) {
             mc.use(mc.std.all.functions[i]);
         }
         mc.useFacade(address(new StdFacade()));
@@ -18,12 +19,12 @@ library DeployLib {
         return mc.deploy(initData).toProxyAddress();
     }
 
-    function deployStdFunctions(MCDevKit storage mc) internal returns(MCDevKit storage) {
+    function deployStdFunctions(MCDevKit storage mc) internal returns (MCDevKit storage) {
         mc.std.functions.deployIfNotExists();
         return mc;
     }
 
-    function deployStdDictionary(MCDevKit storage mc) internal returns(MCDevKit storage) {
+    function deployStdDictionary(MCDevKit storage mc) internal returns (MCDevKit storage) {
         mc.std.complete();
         mc.deployDictionary(mc.std.all);
         return mc;
